Handle error handling on 504 response
Rather than trying to provide a nice error (as we actually do
in find_chandra_obsid when querying OCAT), ditch the idea and just
re-raise it after logging the error. The error handler had erronously
assumed that the reason field would have an errno attribute which
it isn't guaranteed to have.

Rather than try to provide a "nice" reason for the error (at the
expense of losing potentially-useful low-level details) just
re-throw.