Default to python 3.5 rather than python 2.7
The code currently lives in contrib/lib/python<version>/ thanks to not
properly using Python packaging and installation conventions. Rather than
create two packages - one for each version of Python - we have a sym link
from one version to the other (the two supported versions are Python 2.7
and 3.5). Since the plan is to make Pythoon 3.5 the "default" for CIAO
4.10, switch over the symbolic link so that the python 3.5 is the "actual"
version and python 2.7 the linked one.

A number of scripts have code to allow it to be run from within the
development library (again an instance of not using Python packaging
idioms), and they hard code the version is, so they have been updated to
select a version at run time.