Add curl as fallback to deal with https://
cda is going to HTTPS, but CIAO's python doesn't support it (no libssl.so.1)
Need to hack in support to fall back to use curl.

One gotcha is that our limiting sensitivity service only supports
POST requests, not GET, so the URL-encoded hack doesn't work.  limsens
is served from cxc not cda.