Filter PSF map by the exposure map.
Use the MASK functionality in the CIAO Data Model to filter out "unwanted"
pixels in the PSF map. This could have also been achieved with a spatial
filter using the FOV file, but that would require a bit-more work, and this
version should more-accurately match the image.

The idea is to see how well this works.