Initial attempt to handle header merging rules for exptime/expmap
Add a HeaderMerge class to try to emulate the CIAO Data Model header
merging rules. This is specialised for the use case (handling merging
observations) and so may not work in all cases (in particular, those
data with overlapping/ineresting GTIs).

The current implementation is text-based, and should be moved to a
more Python-based version, which could then be used by other parts of
the system (which already deal with creating lookup tables manually).