Add an ardlibqual argument to fullgarf
`fullgarf` runs a loop over all ACIS chips (or HRC plates for the HRC).
When it does, it dynamically constructs the `detsubsys` keyword for `mkgarf`.
Previously, there was no way to add ardlib qualifiers to this `detsubsys`
keyword, it would always use the bare e.g. "ACIS-5".

Unfortunately, that means that fullgarf is incompatible with MARX
simulations, see Chandra-MARX/marx#34 .
This change adds a way to add ardlib qualifiers.