Remove convert_xspec_user_model script
An alternative would be to keep the script but for it to error out
with a "Please contact the HelpDesk" message, but for now just remove
it completely.

Note that there is still support code remaining that is related to
the support of XSPEC user models but I don't want to remove this just
yet in case it is used by other parts of the XSPEC code.