chandra_repro: do not write out a second HISTORY block
The clear acis status bits functionality - called by chandra_repro - added
a HISTORY record in non-standard format, which was causing some confusion.
The change is to make this header addition optional, and for chandra_repro
not to make the change.

One option would be for chandra_repro to call clear_acis_status_bits
with toolname set to 'acis_clear_status_bits', but that is a bit risky
(when/if the script gets removed or renamed).