make_instmap_weights: always set XSPEC abundance and cross-section table
Since the XSPEC defaults can change over time, always set them as there
should be no significant cost to doing so. Note that in XSPEC 12.10.1
the model-library code from XSPEC has changed so that it now picks
a different cross-section (vern rather than bcmc), but we have not
changed the default choice here (the differences are expected to be
small compared to all the other uncertainties in the spectral model
and calibratui uncertainties).