Switch to Sherpa histogram plot support for plot_chart_spectrum
This adds the overplot and clearwindow arguments to plot_chart_spectrum,
and moves the model name from a label (top left of the plot) to the title,
and removes the dataset label.

Labels could be added back, but it requires some backend-specific code
which I am not sure is worth it.