Switch to Sherpa histogram plot support for plot_marx_spectrum
This adds the overplot and clearwindow arguments to plot_marx_spectrum,
and moves the model name from a label (top left of the plot) to the title,
and removes the dataset label.

Labels could be added back, but it requires some backend-specific code
which I am not sure is worth it.

Another change is that the Y axis units of this plot are now the
same as written out to the file (photon/cm^2/s/keV) rather than
that used in the ChaRT plots, namely photon/cm^2/s.

Also includes some minor documentation fixups.