No longer protect underscores in labels
Matplotlib does not treat an underscore as an indicator of a subscript,
so we no-longer need to protect underscores in labels.

This has been noted in the ahelp file for the lc_clean and lc_sigma_clip
routines but not in any other documentation, as it is a small change.