Update instrument-map weight code to make more use of public interfaces
CIAO 4.12 removes the _get_mask method from the Data hierarchy, but prior
to that we can use the mask attribute to get at the values. This change
should not break CIAO 4.11 or 4.12 behavior, but there's a lot of needed
changes to the Sherpa data hierarchy in CIAO 4.12 so there may be other
problems.