Relax keyword requirements for lightcurve (fix regression)
The update to using matplotlib removed a necessary test for accessing
a keyword value, which meant that if the input file didn't have the
necessary keywords the code would now fall over. The code has been
changed to a method that returns None (which was being checked for)
rather than throwing an error.

This should allow light-curve code (deflare and the lc_clean/_sigma_clip
routines) to work with non-Chandra data.