Replace expand_stack with stk.build
The changes here appear more-verbose than they actually are since there
are a number of white-space changes, as well a removal of "old" code
(the setup for testing code), and an un-needed import.

The __version__ field has been removed from these scripts since it is
not used and just another thing that would need to be updated.