Ensure instmap weights can be plotted
The problem is that there is a disconnect between the data plot
preferences (a "curve") and the histogram plot style used here.
This lead to the "histogram" being plotted with a linestyle of
'None' - so nothing was displayed.

Add in support for allowing keyword arguments to be given to
plot_instmap_weights - e.g.

    plot_instmap_weights(ylog=True, color='gray')