check_ciao_version/caldb - work with https site
Clean up the download code for check_ciao_version and check_ciao_caldb
so that it works with the new https location for the CXC. The problem
for these scripts is that the Python urlopen command can fail in
certain platform/system combinations (e.g. ciao-install + Ubuntu)
because the correct certificates can not be found. This was
compounded in the check_ciao_version case by the need to check
multiple URLs, with fall-through behavior on missing files:
this interacted badly with the fall-through-to-wget/curl approach,
resulting in bad results.

The approach now is to use Python only requests - which should be
okay now we can 'import ssl' on the platforms - but to turn off
the TLS validation. This is not ideal, but it is only for
these URL downloads, and they are to cxc.fa.harvard.edu (so no
less safe than it was prior to CIAO 4.12).