check_ciao_version: handle contrib installed manually
This avoids the unhelpful failure message that check_ciao_version
returned if the contrib code has been installed via 'python
setup.py install' (which is only expected to be found for internal
testing).

The code had been written to handle the error case, but I had not
taken advantage of this before. So there's two fixes:

  a) correct handling of the error case
  b) skip checking for any CIAO packages installed via pypi