Object-oriented wrapper for CIAO PSF module
The CIAO psf module gives access to the CALDB REEF file from Python.
However, it's use is a little cumbersome: It requires several lines of
boilterplate code and two imports, just to return one number.
Additionally, the user needs to remember to free recources.

This module warps psf in a OO interface available as a contect manager.