download_chandra_obsid: move from FTP to HTTP(s)
Data is now downloaded via HTTP (well, HTTPS) rather than FTP. The
support for mirrors is retained, but it is assumed that they are also
HTTP(s). FTP support could be added back in, but I'd rather not if at
all possible (and have no information on whether mirrors are available;
I keep the mirror support in as it is useful in case URLs change).

The download order has been changed within an ObsId so that larger
files are downloaded first.