More helpful error for choosing wrong pix_adj parameter
A local user just asked me for help after choosing `pix_adj="edser"` and receiving an "invalid pix_adj" error. I then started to look into the code to see where it came from and only after I realized that the ObsID in question was HRC data. Since the error messages are thrown inside an if block that tests for the instrument name, it can't hurt to print out the instrument name in the error message. When working with just one observation, users should know what they are dealing with, but anyone who processes a number of ObsIDs might not automatically associate ObsID and instrument for all datasets and might appreciate the hint.