Update xspec verison and CIAO version in docs
Really we probably need a variable to update this in automatically on release. Or is it intentional that 4.11 is cited because that's when the tool was added?

TBD: Check if there are other places with the same text, but I'm not doing that until you tell me that we actually *want* to update.

Note: Other models did updates the XSPEC string and CIAO 4.12, e.g. https://cxc.harvard.edu/sherpa/ahelp/xspcfabs.html