Force multiprocessing tasks to use the fork method
The parallel_pool and TaskRunner code now uses an explicit context
when running the code, and it defaults to the 'fork' context.

This means that python 3.8 macOS systems will use fork rather
than spawn.

There is no attempt in this commit to make the option settable by
the user (e.g. by a parameter setting), as I want to see if this
even works before we do any other work.