Adding back convert_xspec_user_model
I have dropped the approach of building via f2py since we needed
to write our own conversion code to create the grids to send to
the XSPEC models. I have instead switched to using Sherpa's
code to do this - following the work I did in
https://github.com/DougBurke/xspeclmodels

This has resulted in a number of changes, such as

 - removal of some of the "fall-through" code to try and
   build code without triggering issues (we'll have to see
   whether the "remove xspec or cfitsio includes" are
   going to need to be added back)

 - the user code (for Sherpa) is now in a {modulename}.ui
   module, as the {modulename} import just loads the models,
   and does not add then to the UI namespace

 - the convolution models now follow the XSPEC convolution
   models (they don't use a load_xxx function)

 - the code is now installed by default, with a --local option
   to build it locally

 - the lmod functionality has been removed as we now can install
   the module globally; it was never a particularly compelling
   addition

 - an error in parsing units was corrected:
   - we no-longer drop spaces between words
   - we handle more cases

 - the module name can be dotted, even if not advertised

The above changes mean that we don't need any support code in
the contributed package to import the module.

I have extended the number of files we have included from the
XSPEC distribution, to pick up both "build" and "install" files.
These are from XSPEC 12.10.1s and so should match CIAO 4.13.
Note that the fftw files (e.g. the FORTRAN module and include
files, as well as the gsl includes, are not included).