Ensure all the XSPEC includes are included
Thanks to confusion with the .gitignore file I had not
included the "build" files. I've renamed the directory
builddir to avoid this issue (rather than fixing it in
the gitignore file).

Note that the

  share/axspec/builddir/tols/raysmith

directory has been manually removed as it is not useful
(and was included because I was matching on \*h)