For *_obs tools, better handle maxsize argument
Issue #262 noted that the tools failed when making PSF images
when using the maxsize argument.

The reason for this is that we don't send in the actual x/y range
to use when creating the exposure maps, but instead re=create the
value using get_sky_limits. For most cases this is okay, because
the limits are "nice" and so the rounding that get_sky_limits does
is not a problem. Unfortunately in the maxsize case this is not
true.

One solution is to send the correct greid through to the
exposure-map code, which is technically the correct thing to do,
and I should do it, but it is invasive.

An alternative (which is still valueable even after we "do the
right thing" above) is to adjust the x/y range that is chosen
when maxsize is set so that it creates a grid we can reliably
access with get_sky_limits. This is the approach taken here,
where we calculate the initial pixel size, and then bump
it up to the next .2/.4/.6/.8 value. So, rather than use a
pixel scale of 4.19140625 we pick 4.2. This has the advantage of
having a half-width of 2.1, and so works well with the output
of get_sky_limits, which defaults to one decimal place.

An alternative would be to just pick the nearest integer value
for the pixel size (or hte next-largest) since is there really
any benefit to using a non-integer value? However, if you have
sub-pixel data you might want something fancy, although in
this case I wuold suggest that the user should set the grid
explicitly rather than rely on maxsize.