Increase number of digits shown in scrflux in screen output
While the correct conf setting is included in the output, it is tempting
for the user to just look at the screen output. However, the screen
output is formatted to show just two digits, which is too few for common
values as as 99.7 %. This PR simply increases the number of digits where
needed without messing up the format where two digits are sufficient:

for conf in [0.1, 0.9, 0.99, 0.9999]:
    print("{:40s}".format( "Value     {:5g}% Conf Interval".format( conf *100)))

Value        10% Conf Interval
Value        90% Conf Interval
Value        99% Conf Interval
Value     99.99% Conf Interval

Note that my editor is set up to automatically strip trailing white space in
Python files. The real change is in line 2433. I can back out the
whitespace changes if needed.