Rework mk_script_tarfile
The argument has now been compressed into a single value

    4.14.0
    4.14.dev

Internally we no-longer use 'python setup.py install' to build the
tarfile because of changes in the python packaging ecosystem - in
particular the install command has been deprecated in setuptools
and instead meant to be handled by pip (or other tool). It looks
like we don't need the logic in setuptools (although donig it
manually has it's own downsides).