Fix up access to obsidmerge header rules
Since we have a fall-through to the dmmerge header lookup rules
the code works, but less-than ideally, when we can no-longer find
the header file. In this case the obsidmerge header keyword list,
which got lost either because code was moved out of bin/ to a
library or because of changes due to packaging.

The fix is simple - send through the locataion of the scripts -
in this case flux_obs and merge_obs - so we can use this. It's not
idal - ideally we'd use importlib.resources/metadata but we are a
long way from that.