When adding images, try to support a large-number of files (#481)
In dmimgcalc_add only, ensure we don't ever try to combine too
many files in a single call, to try to avoid hitting path-length
limits. The current approach is hoaky (it can only support nchunk^2
files at a maximum, where nchunk defaults to 100), but trying to
combine this many files is likely to cause problems elsewhere.

There isn't actually an easy way to determine the maximum number
of files because

- it is likely the operation argument that causes the overflow,
  since the input files are given as a stack
- it is likely the overall command length that is the issue, not
  just a single parameter

Hence the idea to pick 100 as a "nice" chunk size: we hopefully
won't have to do too many chunks for all-but the most large
combinations.

We can only do this for dmimgcalc_add since we have a nice
semigroup here (+). For a generic image combination we can't
know we can do this.