find_mono_energy - tweak the calculation
Ensure the model plot is creating y-axis units of Counts or
Counts/s, and not Counts/keV. This does not matter for standard
Chandra data sets since the energy binning has a constant width
(up to numerical precision) but it could be an issue for
non-Chandra datasets.

There's some safety checks and print-out of internal data
structures to help out debugging.