Use `plot.backend` instead of `backend` to keep up with recent sherpa changes
The version suggested in this PR is backwards compatible and is probably what should have been used all along.
See https://github.com/sherpa/sherpa/pull/1191 for details. Unfortunately, the change in sherpa is necessary to support switching backends and there is no way to keep it working as it was used before this PR.
(OK, it's Python. Ther is always some way, but not a way that's worth it.)

fixes #579