Be explicit with numpy and sherpa build versions
Unfortunately fore us, pip builds the package in a "hermetic" system,
which means it will look in PyPI for a Sherpa version, rather than use
the installed version. This is problematic, as the version on PyPI
could well be a newer version, where code could have changed. So, we
are explicit with the version, to pick up the "same" version as used
in CIAO (or, at least, as close as we can get).

Similarly we go and pick the "compatible" version of NumPy in case
there's an issue. This is slightly different to issue #625 which
suggests using the installed NumPy version. I would hope they are
the same but I can not guarantee it.