Add informative error messages for tools without parameter files
Users (and that includes me before I became part of the team that works on `ciao_contrib`)
don't know and should not have ot care how exactly individual CIAO tools are implemented:
As shell scripts, as C code, as Python code, with or without use or parameter files.

This, users may reasonable expect that `from ciao_contrib import *` gives them access to all the CIAO tools that they find listed on the ahelp pages.
That is not the case and there are good reasons for that (see discussion in https://github.com/cxcsds/ciao-contrib/issues/643#issuecomment-1294078188 .
This PR adds dummys with the names of the missing tools that at the very least will issue a message that the tool is missing and provide some re-direction of what do do instead.

There is the potential downside that now we *do* have e.g. `rt.download_chandra_obsid`, but it dies not do anything. However, I believe it's far more likely (and that's borne out by the questions I've seen so far) that users blindly expect to find all CIAO tools in `runtool` than that they carefully scrutinize the list of module level symbols.

PR includes minor typo fixes in documentation that I reviewed in order to write this code.

closes #643