Take advantage of stack support for dmimgcalc operation parameter
If dmimgcalc is given an opstr that's "too long" (for some
ill-defined criterion of what too long actually is) we switch to
using the stack support, which is new in CIAO 4.15. We could
always do so, but

a) let's not break things we know already work
b) not using a stack here is a bit cleaner

We do not change how dmimgcalc_add works, since even though it
could take advantage of this, we have code that seems to work,
and there may be some advantage of adding data in chunks rather
than all-in-one go, primarily in the amount of memory used at
the expense of header/metadata shenanighans.

There are some minor code-style tweaks.