convert_xspec_user_model: catch XSPEC files
Try to ensure we do not include files created by initpackage - that
is

    lpack_*.cxx
    *FunctionMap.cxx

are not included. We could ensure that the name matches between
these two, but the assumption is that we just drop anything like
these.

There is also some attempt to support case-insensitive systems,
but this is currently untested.