specextract: fix #686
I still don't understand how this bug happens, but I have decided
not to look into the logic behind rstrip and instead write the
code in a simpler manner.

Oh, hold on. It's actualy because rstrip(value) actually treats
value as a list of characters to consider as white space, so
if the outroot parameter ends in one of the characters in "src1",
such as "foo/bar", then the "r" will be included.