combine_spectra: Fix origin of exposure time
The exposure time was taken from the ARF, even when exp_origin='pha' was set (the default),
simple because an ARF exits.
In Chandra, we probably never noticed, but I run into a problem when applying
this script to NICER spectra, where the ARF does not have all the header
information that the PHA would have.
The bug is somewhat obvious in the code and can be tested with any ARF file that does not have an EXPOSURE keyword.
One would get
```
OSError: The file 'NICER_SOMETHING.pi' is missing the 'EXPOSURE' keyword.
```
However, the usual wrapping of the errors in scripts makes it somewhat harder to get to see the details of the error.