Improve URL access (hopefully)
So, if the SSL library doesn't work for your system then we end
up trying an "unverified" SSL context in the URL code. Unfortunately
we did not handle that call in the same way we did to the original
urlopen call, which could lead to the odd behavior we saw in #703.
Even if it didn't, this version should be *slightly* better.