convert_xspec_script: initial support for MDEFINE
This is intended to allow scripts to be created when the
MDEFINE command is used. The script will not be usable, as the
model expression is not converted into Python/Sherpa syntax, but
should allow users to get somewhat further in using XCM files.

Support for adding models is planned, but let's see whether the
changes made here have broken anything.