xcm: Fix the "Fix use of ( rather than )" fix
I had added some internal code "improvements" - such as starting
to add typing annotations - and had messed up some of the run-time
code needed (as I can't for the life of me work out why some
isinstance checks do not seem to get tracked through to calling
code, but maybe that's a design choice of mypy).

I had also done some changes to support allowing mdefine models
to be sent where previously actual XSModel instances were passed,
and the handling of "is this a parameter or not" was borked.