xcm: fixes/improvements for typing
The code takes advantage of the run-time nature of Python a bit
too well, which makes typing accurately a bit of a challenge. It
does point out areas of code that couldbe refactored though!

This also changes the "user model" code to use E for the mid-point
of energy, to better match the XSPEC code.