xcm: a necessary refactoring
Move some of the parsing code out into separate routines where
they so richly belong to be. This allows the typing rules to
"work". There is still a lot of confusion in the code with
names - is "group" /"grouping" a good name to use for XSPEC
"groups". The code could do with a review...