xcm: better handle PHA files with invalid grouping columns
I believe that XSPEC has just "invented" an approach for when
a group of columns has different QUALITY values in it, in that
it seems to add new groups. Sherpa does not do this, treating
the GROUPING column as sacrosanct.

We now identify such issues and allow filter calls to be made
using the XSPEC "channel" numbers (really "group" numbers), and
a warning will be made if it may change the filter used in the
fit.

As part of this the script has switched to usig ui.notice for
the call, rather than calling the notice method on the DataPHA
object, to take advantage of the logging added in CIAO 4.15 to
let the user know what the selected filter is (in the hope of
making things a bit-more obvious).