xcm: add support for the ENERGIES command
This is initial support for the ENERGIES command. Not all features are
supported, and we are not guaranteed to match the XSPEC results (we
extend the response with the user-grid, rather than rebin, as I think
XSPEC does).

Although this uses the underlying technology provided by the
ModelDomainRegridder1D class, the existing interface does not work
reliably for XSPEC models in CIAO 4.16, so we implement our own
version of the code. This will hopefully be removed in a future CIAO
release.