paramio update for CIAO 4.16
We can now access all the parameter information we need with
paramio (even if a bit messy), so switch to that rather than
manually parsing the parameter files.

The results aren't identical, but should be "semantically" - e.g.
numeric values are not guaranteed to have the same textual form
but should have the samve value (e.g. 0.5e-4 vs 5e-5 and 1.0 vs 1).