Usability in aplimits
aplimits can take either bkg_rate or m (number of counts in background region) by setting one of the two to INDEF.
However, m needed to be an int, not INDEF before this PR.
That's not a big problem, because bkg_rate takes precedence (if set, any int number of m will do), but unexpected (no setting for m should be required).
While I'm at it, also fix a few typos in docstrings.