xcm: update notice/ignore handling
This was started because Sherpa 4.17.0 changed the internals of
the DataPHA class, causing the code to fail. In the re-write I
have changed it so it will still work with 4.16.0 but

- will work with 4.17.0
- improves the handling of XSPEC grouping, so we should now understand
  the expected channel range of each XSPEC group, even if Sherpa does
  not group data the same way [*].

One change here is that files with a grouping scheme which differs
between Sherpa and XSPEC will now cause a WARNING message to be
created whenever a xcm.notice call is made. Previously it would only
occur when referencing the "extra group numbers" that XSPEC creates.
The previous behaviour is less correct but probably better for users,
since XCM files are likely to always reference these "extra" bins.

[*] The issue here is what happens if the quality value changes
    between 0 and 2 within a group? XSPEC seems to create a new
    group whereas Sherpa ignores the quality setting and so does
    not create a separate group [!].

[!] Note that XSPEC throws away quality=1 and 5 channels so they
    are never seen whereas Sherpa does not. So XSPEC doesn't have
    to worry about those quality values.