xcm.notice - better handle 4.17.0 filtering
Thanks to changes to etter support CHANNEL values not starting
at 1 we need to rework the notice/ignore handling (which in XSPEC
uses group numbers and not channels).

I had been converting manually to energy or wavelength filters,
but it actually makes sense to keep the filtering in channel
units. There is some work needed to try and replicate the UI
fitering behaviour (that reports the filter change).