xsum: drop ciaotype argument
This used to be stored so we could check if the code had been
copied / installed on a different machine. However, we never
really used it and it's no-longer valid since we do not provide
the information. We could add a track of the platform to the
provenance, but it's not worth it.