Avoid copying the build files.
Go back to the original system where

    pyproject.toml
    meson.build

are added to the top level, and then new code added to

    build_xspec_user_model/

(the Python and C++ code and the meson.build file that compiles
the code and installs things).

This avoids having to copy the source files over, which is needed
as there is no way to guarantee we have identified all the include
files without some clever pre-processing, which is just not worth
it.