versioninfo: minor code style update
Update a comment (due to ciao-install changes) and switch to
using a context manager with urlopen.

There is also a small change to handling the JSON response
which may fix #886 - rather than look for

  "All requested packages already installed."

we just look for a message that starts with

  "All requested packages already installed"

since it seems the presence / absence of a trailing "." is different
with different conda "managers".