Fix issue 883
I assume that something was changed in the 4.15.x series to deal
with non-ASCII PI names which had unforseen consequences for the
error checks. Or something has changed in how a null result is
returned from the CXC Footprint Service. Otherwise I don't see
how I could have not seen this problem.