Start CIAO 4.18 support
As part of this, switch up the build code to use pyproject.toml rather
than setup.cfg. It might be best to drop setuptools completely, but
leave that for later. However, it's not yet clear how well the Python
build infrastructure approach compares to our home-grown code (which
we would like to remove if possible).